fix portability issues and warnings clean and up legacy(#103)
* use C99 variadic macros

* msvc requires the calling convention to be inside the parens

* use windows.h
windef.h is not self-contained

* use _WIN32 for os detection

* fix non-standard construct

* clean up cpuid code
every modern processor has the instruction

* use stdint

* wrap asm blocks in gcc ifdef

* work around msvc crash

* fix variable shadowing

* fix some trivial warnings

* stop generating map files
enable some more warnings

* define missing for msvc

* msvc does not know strcasecmp

* remove obsolete imports

* fix parallel make

* generate faster debug version
NB: Og does enable FPO

* also clean pdbs

* generate pdbs if possible

+++ src/config.c
@@ -1055,7 +1055,7 @@ short calculate_moon_phase(short do_calculate,short add_to_log)
   //Moon phase calculation
   if (do_calculate)
   {
-    phase_of_moon = LbMoonPhase();
+    phase_of_moon = (float)LbMoonPhase();
   }
   if ((phase_of_moon > -0.05) && (phase_of_moon < 0.05))
   {
